Another batch of ui/views narrowing fixes
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3777924
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027071}

+++ ui/views/controls/combobox/combobox.cc
@@ -472,7 +472,7 @@ void Combobox::GetAccessibleNodeData(ui::AXNodeData* node_data) {
     node_data->SetDefaultActionVerb(ax::mojom::DefaultActionVerb::kOpen);
   }
   node_data->AddIntAttribute(ax::mojom::IntAttribute::kPosInSet,
-                             selected_index_.value());
+                             base::checked_cast<int>(selected_index_.value()));
   node_data->AddIntAttribute(ax::mojom::IntAttribute::kSetSize,
                              base::checked_cast<int>(model_->GetItemCount()));
 }
